blockchain/x/iconlake/types/genesis_test.go
package types_test

import (
	"testing"

	"iconlake/x/iconlake/types"

	"github.com/stretchr/testify/require"
)

func TestGenesisState_Validate(t *testing.T) {
	for _, tc := range []struct {
		desc     string
		genState *types.GenesisState
		valid    bool
	}{
		{
			desc:     "default is valid",
			genState: types.DefaultGenesis(),
			valid:    true,
		},
		{
			desc: "valid genesis state",
			genState: &types.GenesisState{

				AccountList: []types.Account{
					{
						AccAddress: []byte("0"),
					},
					{
						AccAddress: []byte("1"),
					},
				},
				// this line is used by starport scaffolding # types/genesis/validField
			},
			valid: true,
		},
		{
			desc: "duplicated account",
			genState: &types.GenesisState{
				AccountList: []types.Account{
					{
						AccAddress: []byte("0"),
					},
					{
						AccAddress: []byte("0"),
					},
				},
			},
			valid: false,
		},
		// this line is used by starport scaffolding # types/genesis/testcase
	} {
		t.Run(tc.desc, func(t *testing.T) {
			err := tc.genState.Validate()
			if tc.valid {
				require.NoError(t, err)
			} else {
				require.Error(t, err)
			}
		})
	}
}
